+++ src/main/java/Moebel/Stuhl.java
@@ -1,7 +1,7 @@
 package Moebel;
 
 public class Stuhl extends Moebel {
-    public Stuhl(String name, int breite, int laenge) {
+    public Stuhl(String name, double breite, double laenge) {
         super(name,breite,laenge);
     }
 }
